[cli] move exceptions to Cli.hpp
No need for a extra header since all modules
include Cli.hpp anyway.
